[ci] unpin workflows from master
We initially pinned workflows to master to help with forwards compatibility (so that old PRs based against master would continue to work).

However, in the long run we want to make the workflows reflect whatever is currently in the local check out, which makes it easier to test changes to workflows.

Since it's been a week since the workflow consolidation: doing the unpinning now. Will follow up with a PR to give a nice message for any PRs that are left stale.

ghstack-source-id: 457163d6f366acb18c3cb9a3c6857b29649ee330
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74817