[ci] add an easier-to-understand error for workflow consolidation FC breaks
If a PR is based against a version of master from before last week (3/21), then our CI will not work as the workflows will depend on files that will not be available in the source tree. We initially mitigated this by retrieving the relevant workflows directly from pytorch master. After we unpin these workflows, stale PRs will start erroring if they are based on an old version of master. This PR adds an easier-to-understand error for those cases.

ghstack-source-id: d094171016665df9230cb2af181f2557de542cf2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74818